Updates download links to local pages
Updates the download links on the main page to point to local HTML files instead of external GitHub release pages.

This allows for more control over the download process and provides a better user experience by keeping users on the site. It also adds the download pages to the git ignore list.